requirements.txt
PyOpenGL==3.1.0a1
PyOpenGL-accelerate==3.1.0a1
numpy==1.8.1
